Split `dbvar` into `delta-store` and `delta-types`
