+++ lib/delta-types/doc/presentation-2021-09-17.pdf
